# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.APIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      prefix:
        type: string
      scopes:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
...
  handlers.CreateAPIKeyRequest:
    properties:
      expires_at:
        type: string
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
    type: object
...
  handlers.CreatedAPIKeyResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      prefix:
        type: string
      scopes:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Accounting app backend API for auth, categories, transactions, stats,
    and API keys.
  title: Sumi API
  version: 0.1.0
paths:
  /api/api-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.APIKeyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      parameters:
      - description: API key payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreatedAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - API Keys
  /api/api-keys/{id}:
    delete:
      parameters:
      - description: API key UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - API Keys
  /api/api-keys/{id}/revoke:
    post:
      parameters:
      - description: API key UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke API key
      tags:
      - API Keys
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Programmatic API key for server-to-server access.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Use the format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"